# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BankManager API",
        "description": "API de gestion bancaire pour BankManager",
        "contact": {
            "email": "contact@bankmanager.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lister tous les clients",
                "description": "Récupère la liste paginée de tous les clients",
                "operationId": "413c12320d4121431e38339f45ecaa7c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Terme de recherche",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des clients récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-clients"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Créer un nouveau client",
                "description": "Crée un nouveau client dans le système",
                "operationId": "f909473745c1a1d469e0c5a5b54682b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-clients"
                        ]
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Afficher un client",
                "description": "Récupère les détails d'un client spécifique",
                "operationId": "feff36e8c89b21277ad216c29ca9ca4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-clients"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Mettre à jour un client",
                "description": "Met à jour les informations d'un client existant",
                "operationId": "4cdaf0a748c704e6f11a98f7bb6eb5a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-clients"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Supprimer un client",
                "description": "Supprime un client du système (soft delete)",
                "operationId": "5b6d3bdaf49885db3b7b4b23aed2156b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID du client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client supprimé avec succès"
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-clients"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client",
                "description": "Modèle représentant un client de la banque",
                "properties": {
                    "id": {
                        "description": "UUID du client",
                        "type": "string",
                        "format": "uuid"
                    },
                    "prenom": {
                        "description": "Prénom du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "nom": {
                        "description": "Nom du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "nom_complet": {
                        "description": "Nom complet du client",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email du client",
                        "type": "string",
                        "format": "email"
                    },
                    "telephone": {
                        "description": "Téléphone du client",
                        "type": "string",
                        "maxLength": 20,
                        "nullable": true
                    },
                    "adresse": {
                        "description": "Adresse du client",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "ville": {
                        "description": "Ville du client",
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "pays": {
                        "description": "Pays du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "code_postal": {
                        "description": "Code postal",
                        "type": "string",
                        "maxLength": 10,
                        "nullable": true
                    },
                    "numero_identification": {
                        "description": "Numéro d'identification (CIN, Passeport, etc.)",
                        "type": "string"
                    },
                    "type_identification": {
                        "description": "Type d'identification",
                        "type": "string",
                        "enum": [
                            "CIN",
                            "Passeport",
                            "Permis de conduire"
                        ]
                    },
                    "date_naissance": {
                        "description": "Date de naissance",
                        "type": "string",
                        "format": "date"
                    },
                    "age": {
                        "description": "Âge du client",
                        "type": "integer"
                    },
                    "sexe": {
                        "description": "Sexe du client",
                        "type": "string",
                        "enum": [
                            "M",
                            "F",
                            "Autre"
                        ]
                    },
                    "profession": {
                        "description": "Profession du client",
                        "type": "string",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "employeur": {
                        "description": "Employeur du client",
                        "type": "string",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "revenu_mensuel": {
                        "description": "Revenu mensuel",
                        "type": "number",
                        "format": "decimal",
                        "nullable": true
                    },
                    "statut": {
                        "description": "Statut du client",
                        "type": "string",
                        "enum": [
                            "Actif",
                            "Inactif",
                            "Suspendu",
                            "Bloqué"
                        ]
                    },
                    "notes": {
                        "description": "Notes sur le client",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "datetime"
                    },
                    "comptes": {
                        "description": "Comptes du client",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Compte"
                        }
                    }
                },
                "type": "object"
            },
            "Compte": {
                "title": "Compte",
                "description": "Modèle représentant un compte bancaire",
                "properties": {
                    "id": {
                        "description": "UUID du compte",
                        "type": "string",
                        "format": "uuid"
                    },
                    "numero_compte": {
                        "description": "Numéro du compte",
                        "type": "string",
                        "maxLength": 20
                    },
                    "titulaire": {
                        "description": "Nom du titulaire",
                        "type": "string",
                        "maxLength": 200
                    },
                    "type": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ]
                    },
                    "solde": {
                        "description": "Solde du compte",
                        "type": "number",
                        "format": "decimal"
                    },
                    "solde_formate": {
                        "description": "Solde formaté avec devise",
                        "type": "string"
                    },
                    "devise": {
                        "description": "Devise du compte",
                        "type": "string",
                        "maxLength": 3
                    },
                    "date_creation": {
                        "description": "Date de création du compte",
                        "type": "string",
                        "format": "date"
                    },
                    "statut": {
                        "description": "Statut du compte",
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ]
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "datetime"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    },
                    "transactions": {
                        "description": "Transactions du compte",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    }
                },
                "type": "object"
            },
...
            "StoreClientRequest": {
                "title": "StoreClientRequest",
                "description": "Données pour créer un nouveau client",
                "required": [
                    "prenom",
                    "nom",
                    "email",
                    "numero_identification",
                    "type_identification",
                    "date_naissance",
                    "sexe",
                    "pays"
                ],
                "properties": {
                    "prenom": {
                        "description": "Prénom du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "nom": {
                        "description": "Nom du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "description": "Email du client",
                        "type": "string",
                        "format": "email"
                    },
                    "telephone": {
                        "description": "Téléphone du client",
                        "type": "string",
                        "maxLength": 20,
                        "nullable": true
                    },
                    "adresse": {
                        "description": "Adresse du client",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "ville": {
                        "description": "Ville du client",
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "pays": {
                        "description": "Pays du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "code_postal": {
                        "description": "Code postal",
                        "type": "string",
                        "maxLength": 10,
                        "nullable": true
                    },
                    "numero_identification": {
                        "description": "Numéro d'identification",
                        "type": "string"
                    },
                    "type_identification": {
                        "description": "Type d'identification",
                        "type": "string",
                        "enum": [
                            "CIN",
                            "Passeport",
                            "Permis de conduire"
                        ]
                    },
                    "date_naissance": {
                        "description": "Date de naissance",
                        "type": "string",
                        "format": "date"
                    },
                    "sexe": {
                        "description": "Sexe du client",
                        "type": "string",
                        "enum": [
                            "M",
                            "F",
                            "Autre"
                        ]
                    },
                    "profession": {
                        "description": "Profession du client",
                        "type": "string",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "employeur": {
                        "description": "Employeur du client",
                        "type": "string",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "revenu_mensuel": {
                        "description": "Revenu mensuel",
                        "type": "number",
                        "format": "decimal",
                        "nullable": true
                    },
                    "notes": {
                        "description": "Notes sur le client",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "title": "PaginationLinks",
                "description": "Liens de pagination",
                "properties": {
                    "first": {
                        "description": "Lien vers la première page",
                        "type": "string"
                    },
                    "last": {
                        "description": "Lien vers la dernière page",
                        "type": "string"
                    },
                    "prev": {
                        "description": "Lien vers la page précédente",
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "description": "Lien vers la page suivante",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Métadonnées de pagination",
                "properties": {
                    "current_page": {
                        "description": "Page actuelle",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Premier élément de la page",
                        "type": "integer"
                    },
                    "last_page": {
                        "description": "Dernière page",
                        "type": "integer"
                    },
                    "links": {
                        "description": "Liens de pagination",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "path": {
                        "description": "URL de base",
                        "type": "string"
                    },
                    "per_page": {
                        "description": "Éléments par page",
                        "type": "integer"
                    },
                    "to": {
                        "description": "Dernier élément de la page",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Nombre total d'éléments",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel Passport OAuth2 security.",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/oauth/token/refresh",
                        "scopes": {
                            "read-clients": "Lire les clients",
                            "write-clients": "Créer/modifier les clients",
                            "read-comptes": "Lire les comptes",
                            "write-comptes": "Créer/modifier les comptes",
                            "read-transactions": "Lire les transactions",
                            "write-transactions": "Créer des transactions"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": [
                "read-clients",
                "write-clients",
                "read-comptes",
                "write-comptes",
                "read-transactions",
                "write-transactions"
            ]
        }
    ]
}